added new file dispatecher file

+++ src/main/java/com/nextech/erp/model/Rawmaterial.java
@@ -26,7 +26,7 @@ public class Rawmaterial implements Serializable {
 	private long id;
 
 	@Column(name="created_by")
-	private int createdBy;
+	private long createdBy;
 
 	@Column(name="created_date")
 	private Timestamp createdDate;
@@ -44,7 +44,7 @@ public class Rawmaterial implements Serializable {
 	private float pricePerUnit;
 
 	@Column(name="updated_by")
-	private int updatedBy;
+	private long updatedBy;
 
 	@Column(name="updated_date")
 	private Timestamp updatedDate;
@@ -97,11 +97,11 @@ public class Rawmaterial implements Serializable {
 		this.id = id;
 	}
 
-	public int getCreatedBy() {
+	public long getCreatedBy() {
 		return this.createdBy;
 	}
 
-	public void setCreatedBy(int createdBy) {
+	public void setCreatedBy(long createdBy) {
 		this.createdBy = createdBy;
 	}
 
@@ -153,11 +153,11 @@ public class Rawmaterial implements Serializable {
 		this.pricePerUnit = pricePerUnit;
 	}
 
-	public int getUpdatedBy() {
+	public long getUpdatedBy() {
 		return this.updatedBy;
 	}
 
-	public void setUpdatedBy(int updatedBy) {
+	public void setUpdatedBy(long updatedBy) {
 		this.updatedBy = updatedBy;
 	}
